+++ packages/expo-router/src/layouts/StackClient.tsx
@@ -22,6 +22,7 @@ import { StackAnimationTypes } from 'react-native-screens';
 import { withLayoutContext } from './withLayoutContext';
 import { createNativeStackNavigator } from '../fork/native-stack/createNativeStackNavigator';
 import { useLinkPreviewContext } from '../link/preview/LinkPreviewContext';
+import { getInternalExpoRouterParams, type InternalExpoRouterParams } from '../navigationParams';
 import { SingularOptions, getSingularId } from '../useScreens';
 import { Protected } from '../views/Protected';
 
@@ -79,10 +80,7 @@ type RNNavigationAction = Extract<CommonNavigationAction, { type: 'NAVIGATE' }>;
 type RNPreloadAction = Extract<CommonNavigationAction, { type: 'PRELOAD' }>;
 type ExpoNavigationAction = Omit<RNNavigationAction, 'payload'> & {
   payload: Omit<RNNavigationAction['payload'], 'params'> & {
-    params: {
-      __internal__expoRouterIsPreviewNavigation?: boolean;
-      params?: Record<string, unknown>;
-    };
+    params: RNNavigationAction['payload']['params'] & InternalExpoRouterParams;
   };
 };
 
@@ -102,10 +100,10 @@ function isStackAction(
 const isPreviewAction = (action: NavigationAction): action is ExpoNavigationAction =>
   !!action.payload &&
   'params' in action.payload &&
-  !!action.payload.params &&
-  typeof action.payload === 'object' &&
-  '__internal__expoRouterIsPreviewNavigation' in (action.payload.params as any) &&
-  !!(action.payload.params as any).__internal__expoRouterIsPreviewNavigation;
+  typeof action.payload.params === 'object' &&
+  !!getInternalExpoRouterParams(action.payload?.params ?? undefined)[
+    '__internal__expo_router_is_preview_navigation'
+  ];
 
 /**
  * React Navigation matches a screen by its name or a 'getID' function that uniquely identifies a screen.
@@ -191,7 +189,7 @@ export const stackRouterOverride: NonNullable<ComponentProps<typeof RNStack>['UN
           }
 
           // START FORK
-          if (isPreviewAction(action)) {
+          if (isPreviewAction(action) && !route) {
             route = state.preloadedRoutes.find(
               (route) => route.name === action.payload.name && id === route.key
             );