+++ packages/expo-router/CHANGELOG.md
@@ -11,6 +11,7 @@
 ### 💡 Others
 
 - Throw an error when tabs are nested ([#38651](https://github.com/expo/expo/pull/38651) by [@Ubax](https://github.com/Ubax))
+- refactor internal params logic ([#38929](https://github.com/expo/expo/pull/38929) by [@Ubax](https://github.com/Ubax))
 
 ## 6.0.0-beta.2 — 2025-08-16